1vapigw-k8s-master/utils/active-cluster-check/requirements.txt
requests~=2.31.1
PyGithub~=1.55
